Move recursive union entailment checks to integration
