+++ tools/ts2pant/tests/integration/recursive-union.test.mts
@@ -23,11 +23,20 @@ describe("recursive discriminated union integration", () => {
     await loadAst();
   });
 
-  it("a @pant annotation over a recursive union entails", {
-    skip: !solverAvailable() ? "z3 not available" : undefined,
-  }, async () => {
-    const output = await emitAndCheck(await buildDocument(FIXTURE, "treeKind"));
-    const result = runCheck(output);
-    assert.ok(result.passed, `pant --check failed:\n${result.output}`);
-  });
+  for (const fn of ["treeKind", "leafValue", "leftChildKind"]) {
+    it(`@pant on \`${fn}\` (recursive union) entails`, {
+      skip: !solverAvailable() ? "z3 not available" : undefined,
+    }, async () => {
+      const output = await emitAndCheck(await buildDocument(FIXTURE, fn));
+      const result = runCheck(output);
+      const entailments = result.checks.filter(
+        (c) =>
+          c.message.startsWith("OK: Entailed:") ||
+          c.message.startsWith("FAIL: Not entailed:"),
+      );
+      assert.ok(entailments.length >= 1, "expected an entailment result");
+      assert.ok(result.passed, `pant --check failed:\n${result.output}`);
+      assert.ok(entailments.every((c) => c.passed));
+    });
+  }
 });
