+++ tools/ts2pant/tests/recursive-union.test.mts
@@ -5,15 +5,9 @@ import assert from "node:assert/strict";
 import { resolve } from "node:path";
 import { before, describe, it } from "node:test";
 
-import { emitDocument, runCheck } from "../src/emit.js";
+import { emitDocument } from "../src/emit.js";
 import { loadAst } from "../src/pant-wasm.js";
-import {
-  buildDocument,
-  emitAndCheck,
-  getPantBin,
-  PROJECT_ROOT,
-  solverAvailable,
-} from "./helpers.mjs";
+import { buildDocument, emitAndCheck } from "./helpers.mjs";
 
 const FIXTURE = resolve(
   import.meta.dirname,
@@ -46,22 +40,8 @@ describe("recursive discriminated union", () => {
   // `node.left: Tree` field under `kind === "node"` narrowing, then its
   // discriminant — exercising the recursive accessor `tree--left … => Tree`.
   for (const fn of ["treeKind", "leafValue", "leftChildKind"]) {
-    it(`@pant on \`${fn}\` (recursive union) entails`, {
-      skip: !solverAvailable() ? "z3 not available" : undefined,
-    }, async () => {
-      const output = await emitAndCheck(await buildDocument(FIXTURE, fn));
-      const result = runCheck(output, {
-        projectRoot: PROJECT_ROOT,
-        pantBin: getPantBin(),
-      });
-      const entailments = result.checks.filter(
-        (c) =>
-          c.message.startsWith("OK: Entailed:") ||
-          c.message.startsWith("FAIL: Not entailed:"),
-      );
-      assert.ok(entailments.length >= 1, "expected an entailment result");
-      assert.ok(result.passed, `pant --check failed:\n${result.output}`);
-      assert.ok(entailments.every((c) => c.passed));
+    it(`emits and type-checks @pant on \`${fn}\` (recursive union)`, async () => {
+      await emitAndCheck(await buildDocument(FIXTURE, fn));
     });
   }
